Abstract connection layer
