+++ server.go
@@ -1,113 +1,10 @@
 package gonameko
 
-import (
-	"encoding/json"
-	"fmt"
-	"log"
-
-	"github.com/streadway/amqp"
-)
-
 type Server struct {
-	Name           string
-	RabbitHostname string
-	RabbitUser     string
-	RabbitPass     string
-	RabbitPort     int64
-	ContentType    string
-
-	conn    *amqp.Connection
-	channel *amqp.Channel
-	queue   *amqp.Queue
+	Name string
+	Conn *Connection
 }
 
 func (s *Server) Run() {
-	url := fmt.Sprintf("amqp://%v:%v@%v:%v/", s.RabbitUser, s.RabbitPass, s.RabbitHostname, s.RabbitPort)
-	conn, err := amqp.Dial(url)
-	FailOnError(err, "Failed to connect to RabbitMQ")
-	s.conn = conn
-
-	ch, err := conn.Channel()
-	FailOnError(err, "Failed to open a channel")
-	s.channel = ch
-
-	err = ch.ExchangeDeclare(
-		"nameko-rpc", // name
-		"topic",      // type
-		true,         // durable
-		false,        // auto-deleted
-		false,        // internal
-		false,        // no-wait
-		nil,          // arguments
-	)
-	FailOnError(err, "Failed to declare an exchange")
-
-	q, err := ch.QueueDeclare(
-		fmt.Sprintf("rpc-%v", s.Name), // name
-		false,                         // durable
-		false,                         // delete when unused
-		false,                         // exclusive
-		false,                         // no-wait
-		nil,                           // arguments
-	)
-	FailOnError(err, "Failed to declare a queue")
-	s.queue = &q
-
-	err = ch.Qos(
-		10,    // prefetch count
-		0,     // prefetch size
-		false, // global
-	)
-	FailOnError(err, "Failed to set QoS")
-
-	err = ch.QueueBind(
-		q.Name,
-		fmt.Sprintf("%v.*", s.Name),
-		"nameko-rpc",
-		false,
-		nil,
-	)
-	FailOnError(err, "Failed to bind a queue")
-
-	msgs, err := ch.Consume(
-		s.queue.Name, // queue
-		"",           // consumer
-		false,        // auto ack
-		false,        // exclusive
-		false,        // no local
-		false,        // no wait
-		nil,          // args
-	)
-	FailOnError(err, "Failed to register a consumer")
-
-	forever := make(chan bool)
-
-	go func() {
-		for msg := range msgs {
-			fmt.Println(msg)
-			response, _ := json.Marshal(
-				RPCResponse{
-					Result: "hello, nameko!",
-					Err:    nil,
-				},
-			)
-
-			err := s.channel.Publish(
-				"nameko-rpc",
-				msg.ReplyTo,
-				false,
-				false,
-				amqp.Publishing{
-					ContentType:   "application/xjson",
-					CorrelationId: msg.CorrelationId,
-					Body:          response,
-				})
-			FailOnError(err, "Failed to publish a message")
-
-			msg.Ack(false)
-		}
-	}()
-
-	log.Printf(" [*] Server is waiting...")
-	<-forever
+	s.Conn.Serve(s.Name)
 }
